+++ release_notes.md
@@ -10,6 +10,7 @@ this work is available on `master` branch, but has not yet been released.
   * support for tests in `catch2` inside the library code
   * `But::System::SocketPair` added (RAII-wrapper for `socketpair()` syscall)
   * `But::System::syscallRetry` added wrapper that auto-restarts syscall that failed spuriously
+  * `But::System::makeNonblocking` wrapper that makes a socket non-blocking
 
 
 ## v2.0.0 (2023-08-13)